+++ PACECloudSDK/AppKit/API/MessageHandling/ScriptMessageHandler.swift
@@ -1,0 +1,13 @@
+//
+//  ScriptMessageHandler.swift
+//  PACECloudSDK
+//
+//  Created by PACE Telematics GmbH.
+//
+
+import Foundation
+
+enum ScriptMessageHandler: String, CaseIterable {
+    case nativeAPI = "pace_native_api"
+    case logger = "pace_logger"
+}